pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devsecops-agent"
version = "0.1.0"
description = "CLI foundation for a DevSecOps scanning tool."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Praveen Jakku" }
]
dependencies = [
  "PyYAML>=6.0.1,<7.0.0",
  "typer>=0.12.3,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[project.scripts]
devsecops-agent = "devsecops_agent.cli:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["devsecops_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
